% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_population.R
\name{fish_population}
\alias{fish_population}
\title{Fish the Population}
\usage{
fish_population(fishArea, location, max.prob, delta.prob)
}
\arguments{
\item{location}{vector specifying row and column to fish in}

\item{max.prob}{probability of catching fish in specified cells}

\item{delta.prob}{change in probability getting further from specified location}

\item{fish.area}{output from initialize_population function}
}
\description{
Function to fish the population
}

