% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_fish_cw.R
\name{move_fish_cw}
\alias{move_fish_cw}
\title{Move Fish Clockwise}
\usage{
move_fish_cw(fish_area, move_prob, seed = 300, ...)
}
\arguments{
\item{fish_area}{Input matrix of distributed fish}

\item{move_prob}{probability of moving from one quadrant to the next}
}
\description{
This function is meant to simulate fishing moving offshore. One function call results in 
fish moving one column to the left with some probability (specified in the arguments). 
Probabilities decline evenly by column (highest on furthest left column).
}
\examples{

}
\keyword{movement}

