% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_fish_hs.R
\name{move_fish_hs}
\alias{move_fish_hs}
\title{Move Fish based on Hotspots}
\usage{
move_fish_hs(fish_area, hotspots, max_prob = 0.5, ...)
}
\arguments{
\item{fish_area}{Input matrix of distributed fish. Use initialize population to specify this}

\item{hotspots}{Specify the location of hotspots as a data frame with the names}

\item{max_prob}{Maximum probability of moving toward the hotspots. Cells adjacent to 
the hotspots have the highest probability, cells two cells from the hotspots have a probability
of moving that is half max_prob.}
}
\description{
This function is meant to simulate fish moving towards hotspots. The idea is that rockfish
tend to aggregate around rocky structure. The user specifies hotspots and the
fish then move towards the hotspots with some probability max_prob.
}
\examples{
move_fish_hs(fish_area = initialize_population(distribute = 'patchy', numrow = 10, numcol = 10, seed = 5,
                         nfish = 1000, percent = .50), 
             hotspots = data.frame('row' = c(3, 2), 'column' = c(3, 5)),
             max_prob = 0.2)
}
\keyword{movement}

