% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_fish_left.R
\name{move_fish_left}
\alias{move_fish_left}
\title{Move Fish Left}
\usage{
move_fish_left(fish_area, max_prob = 0.5, min_prob = 0.05, ...)
}
\arguments{
\item{fish_area}{Input matrix of distributed fish}

\item{max_prob}{maximum movement probability in the furthest left location}

\item{min_prob}{minimum movement probability in furthest right location}
}
\description{
This function is meant to simulate fishing moving offshore. One function call results in 
fish moving one column to the left with some probability (specified in the arguments). 
Probabilities decline evenly by column (highest on furthest left column).
}
\examples{

}
\keyword{movement}

