% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_fish_none.R
\name{move_fish_none}
\alias{move_fish_none}
\title{Move Fish None}
\usage{
move_fish_none(fish_area, ...)
}
\arguments{
\item{fish_area}{Input matrix of distributed fish}
}
\description{
This function doesn't move any fish. Easiest to add no movement into the package
with this dummy function
}
\examples{

}
\keyword{movement}

