% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_master_list.R
\name{parse_master_list}
\alias{parse_master_list}
\title{parse_master_list}
\usage{
parse_master_list(master_list)
}
\arguments{
\item{master_list}{output from survey_over_years}
}
\description{
This function initializes the spatial distribution of the fish population
}
\examples{
xx <- survey_over_years(numrow = 10, numcol = 10, nfish = 10000, 
  distribute = 'uniform',
  seed = 300, nyears = 15, location_list, 
  random_locations = TRUE, nlocs = 10, move_func = move_fish_cw, move_prob = .8)
thing <- parse_master_list(xx)
}
\keyword{initialize}

