% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annual_cpue_nfish.R
\name{plot_annual_cpue_nfish}
\alias{plot_annual_cpue_nfish}
\title{Plot annual cpue nfish}
\usage{
plot_annual_cpue_nfish(parsed_list, cex.axis = 1.2, text_space = 0.25,
  print_text = TRUE, ...)
}
\arguments{
\item{parsed_list}{output of survey_over_years melted for formatting}

\item{cex.axis}{spcify size of axis, all other cex paramters scaled to this}

\item{text_space}{spacing inbetween text descriptions in upper left of plot}
}
\description{
Plot annual averaged cpue and true number of fish
}
\examples{
plot example
}
\keyword{plot}

