% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hook_prob_equal.R, R/sample_equal_prob.R
\name{sample_equal_prob}
\alias{sample_equal_prob}
\title{Conduct Survey
Wrapper to repeatedly run fish_population function
Give function arguments for
This function initializes the spatial distribution of the fish population}
\usage{
sample_equal_prob(nfish = 100, nhooks = 5, p0 = 0.001, samp_size = 3)

sample_equal_prob(nfish = 100, nhooks = 5, p0 = 0.001, samp_size = 3)
}
\arguments{
\item{nfish}{number of fish, probability calcuations are usually to the power nfish}

\item{nhooks}{number of hooks}

\item{p0}{probability that each fish detects gear}

\item{p0}{probability that each fish detects gear}

\item{nfish}{number of fish, probability calcuations are usually to the power nfish}

\item{nhooks}{number of hooks}
}
\description{
Conduct Survey
Wrapper to repeatedly run fish_population function
Give function arguments for
This function initializes the spatial distribution of the fish population

Conduct Survey
Wrapper to repeatedly run fish_population function
Give function arguments for
This function initializes the spatial distribution of the fish population
}
\examples{
put example here dude

put example here dude

}
\keyword{hook}
\keyword{probability}

