% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_over_years.R
\name{survey_over_years}
\alias{survey_over_years}
\title{Initialize Population}
\usage{
survey_over_years(numrow = 10, numcol = 10, nfish = 1000, distribute,
  seed = 300, nyears = 15, location_list, random_locations = FALSE,
  nlocs = 10, move_func, nhooks, ndrops, scope, ...)
}
\arguments{
\item{numrow}{Number of rows in matrix}

\item{numcol}{Number of columns in matrix}

\item{nfish}{Number of fish to allocate among matrix}

\item{distribute}{Specify fish distribution to be 'uniform', 'patchy', or 'area' specific}

\item{seed}{Set seed if distribute == random, defaults to 300}

\item{maxfish}{Maximum number of fish that can be sampled at a time}

\item{percent}{percentage of area to sample. Only necessary if distribute == 'patchy'}

\item{area}{Specify area to distribute fish, options are 'upperleft', 'upperright', 'lowerleft', 'lowerright',
'upperhalf', 'lowerhalf', 'righthalf', 'lefthalf' 
  Only necessary if distribute == 'area'}
}
\description{
This function initializes the spatial distribution of the fish population
}
\examples{

Uniformly distribute fish
initialize_population(numrow = 10, numcol = 10, nfish = 1000, distribute = 'uniform')

Distribute fish in upper right quadrant only
initialize_population(numrow = 10, numcol = 10, nfish = 1000, distribute = 'area',
                            area = 'upperright')

Distribute fish in upper half
initialize_population(numrow = 10, numcol = 10, nfish = 1000, distribute = 'area',
                            area = 'upperhalf')

Patchily distribute fish
initialize_population(numrow = 10, numcol = 10, nfish = 100, distribute = 'patchy',
  percent = .2)


}
\keyword{initialize}

