% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_population.R
\name{initialize_population}
\alias{initialize_population}
\title{Initialize Population}
\usage{
initialize_population(nrow, ncol, nfish = 100, distribute, seed = 300)
}
\arguments{
\item{nrow}{Number of rows in matrix}

\item{ncol}{Number of columns in matrix}

\item{nfish}{Number of fish to allocate among matrix}

\item{distribute}{should distribution be uniform or random}

\item{seed}{Set seed if distribute == random, defaults to 300}
}
\description{
This function initializes the spatial distribution of the fish population
}
\examples{
put example here dude
}
\keyword{initialize}

