% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_population.R
\name{initialize_population}
\alias{initialize_population}
\title{Initialize Population}
\usage{
initialize_population(numrow, numcol, nfish = 100, distribute, seed = 300,
  area = "upperleft")
}
\arguments{
\item{numrow}{Number of rows in matrix}

\item{numcol}{Number of columns in matrix}

\item{nfish}{Number of fish to allocate among matrix}

\item{distribute}{Specify if the fish distribution be random or uniform}

\item{seed}{Set seed if distribute == random, defaults to 300}

\item{area}{Specify area to distribute fish, options are 'upperleft', 'upperright', 'lowerleft', 'lowerright'}
}
\description{
This function initializes the spatial distribution of the fish population
}
\examples{
put example here dude
}
\keyword{initialize}

