% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_population.R
\name{initialize_population}
\alias{initialize_population}
\title{Initialize Population}
\usage{
initialize_population(numrow, numcol, nfish = 100, seed = 300, distribute,
  maxfish = 10, area = "upperleft", percent = 0.1)
}
\arguments{
\item{numrow}{Number of rows in matrix}

\item{numcol}{Number of columns in matrix}

\item{nfish}{Number of fish to allocate among matrix}

\item{seed}{Set seed if distribute == random, defaults to 300}

\item{distribute}{Specify if the fish distribution be 'uniform', 'patchy' or 'random'}

\item{maxfish}{Maximum number of fish that can be sampled at a time}

\item{area}{Specify area to distribute fish, options are 'upperleft', 'upperright', 'lowerleft', 'lowerright'. 
Only necessary if distribute == 'random'}

\item{percent}{percentage of area to sample. Only necessary if distribute == 'patchy'}
}
\description{
This function initializes the spatial distribution of the fish population
}
\examples{

initialize_population(numrow = 10, numcol = 10, nfish = 1000, distribute = 'uniform')

initialize_population(numrow = 10, numcol = 10, nfish = 1000, distribute = 'area',
                            area = 'upperright')

initialize_population(numrow = 10, numcol = 10, nfish = 100, distribute = 'patchy',
  percent = .2)
}
\keyword{initialize}

